// src/app/swagger/model/orderDto.ts
/**
 * Backend
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export interface OrderDto { 
    id?: number;
    inputDate?: string;
    deliveryDate?: string | null;
    billCreatedDate?: string | null;
    paymentDate?: string | null;
    documentNr?: number;
    brutto?: number;
    netto?: number;
    tax?: number;
    customerName?: string | null;
    bill?: number | null;
    po?: string | null;
    status?: string | null;
    uid?: string | null;
}